BrightLite Separation of Configs
Some backend work for readability since the code will grow more complex with Nightlite in the same PBO.
